requirements.txt
streamlit

langchain==0.2.17
langchain-community==0.2.19
langchain-chroma==0.1.2
langchain-groq==0.1.10
langchain-huggingface==0.0.3

chromadb==0.4.24
pysqlite3-binary        # ← ADD THIS LINE — fixes SQLite on Streamlit Cloud

sentence-transformers==2.7.0
transformers==4.41.2
torch==2.2.2

python-dotenv
pypdf
protobuf==3.20.3